finalize update artwork popup
